# docs/swagger.yaml
basePath: /
definitions:
...
  genproto.CategoryDeleteResponse:
    properties:
      success:
        type: boolean
    type: object
  genproto.CategoryResponse:
    properties:
      category_id:
        type: string
      name:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.CreateCategoryRequest:
    properties:
      id:
        type: string
      name:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  genproto.CreateGoalRequest:
    properties:
      current_amount:
        type: number
      deadline:
        type: string
      id:
        type: string
      status:
        type: string
      target_amount:
        type: number
      user_id:
        type: string
    type: object
...
  genproto.GoalDeleteResponse:
    properties:
      success:
        type: boolean
    type: object
  genproto.GoalResponse:
    properties:
      current_amount:
        type: number
      deadline:
        type: string
      goal_id:
        type: string
      target_amount:
        type: number
      user_id:
        type: string
    type: object
...
  genproto.ListGoalsResponse:
    properties:
      goals:
        items:
          $ref: '#/definitions/genproto.GoalResponse'
        type: array
    type: object
  genproto.ListResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/genproto.CategoryResponse'
        type: array
    type: object
  genproto.MessageResponse:
    properties:
      Message:
        type: string
    type: object
...
  genproto.Responsee:
    properties:
      Message:
        type: string
    type: object
...
  genproto.UpdateCategoryRequest:
    properties:
      category_id:
        type: string
      name:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  genproto.UpdateGoalRequest:
    properties:
      courent_amount:
        type: number
      deadline:
        type: string
      goal_id:
        type: string
      target_amount:
        type: number
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Auth service API documentation
  title: api gat way
  version: "1.0"
paths:
...
  /category/create:
    post:
      consumes:
      - application/json
      description: Create a new category for a user
      parameters:
      - description: Create category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category created successfully
          schema:
            $ref: '#/definitions/genproto.MessageResponse'
        "500":
          description: Error while creating category
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Category
      tags:
      - Category
  /category/delete/{category_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing category by its ID
      parameters:
      - description: Category ID
        in: query
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            $ref: '#/definitions/genproto.CategoryDeleteResponse'
        "400":
          description: 'Missing required query parameter: category_id'
          schema:
            type: string
        "500":
          description: Error while deleting category
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Category
      tags:
      - Category
  /category/get/{category_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a category's details by its ID
      parameters:
      - description: Category ID
        in: query
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category details
          schema:
            $ref: '#/definitions/genproto.CategoryResponse'
        "400":
          description: 'Missing required query parameter: category_id'
          schema:
            type: string
        "500":
          description: Error while fetching category
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Category by ID
      tags:
      - Category
  /category/list:
    get:
      consumes:
      - application/json
      description: List all categories for a user
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            $ref: '#/definitions/genproto.ListResponse'
        "500":
          description: Error while listing categories
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List Categories
      tags:
      - Category
  /category/update:
    put:
      consumes:
      - application/json
      description: Update the details of an existing category
      parameters:
      - description: Update category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
          schema:
            $ref: '#/definitions/genproto.MessageResponse'
        "400":
          description: Invalid request or missing category_id
          schema:
            type: string
        "500":
          description: Error while updating category
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Category
      tags:
      - Category
  /goal/create:
    post:
      consumes:
      - application/json
      description: Create a new financial goal for a user
      parameters:
      - description: Create goal request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateGoalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Goal created successfully
          schema:
            $ref: '#/definitions/genproto.Responsee'
        "500":
          description: Error while creating goal
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Goal
      tags:
      - Goal
  /goal/delete/{goal_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing goal by its ID
      parameters:
      - description: Goal ID
        in: query
        name: goal_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Goal deleted successfully
          schema:
            $ref: '#/definitions/genproto.GoalDeleteResponse'
        "400":
          description: 'Missing required query parameter: goal_id'
          schema:
            type: string
        "500":
          description: Error while deleting goal
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Goal
      tags:
      - Goal
  /goal/get/{goal_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a goal's details by its ID
      parameters:
      - description: Goal ID
        in: query
        name: goal_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Goal details
          schema:
            $ref: '#/definitions/genproto.GoalResponse'
        "400":
          description: 'Missing required query parameter: goal_id'
          schema:
            type: string
        "500":
          description: Error while fetching goal
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Goal by ID
      tags:
      - Goal
  /goal/list:
    get:
      consumes:
      - application/json
      description: List all financial goals for a user
      produces:
      - application/json
      responses:
        "200":
          description: List of goals
          schema:
            $ref: '#/definitions/genproto.ListGoalsResponse'
        "500":
          description: Error while listing goals
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List Goals
      tags:
      - Goal
  /goal/update:
    put:
      consumes:
      - application/json
      description: Update the details of an existing goal
      parameters:
      - description: Update goal request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateGoalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Goal updated successfully
          schema:
            $ref: '#/definitions/genproto.Responsee'
        "400":
          description: Invalid request or missing goal_id
          schema:
            type: string
        "500":
          description: Error while updating goal
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Goal
      tags:
      - Goal
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"